feat: Create index in OpenSearch
This commit creates enables the lambda function to create an index
in OpenSearch. Current approach requires the user to provide his
secret access key and access key id.
This is not secure and should be changed in the future.

Additionally, creating the stacks now have unique names to avoid based
on stage name duplication leading to errors.